+++ .storybook/preview.tsx
@@ -128,6 +128,7 @@ const preview: Preview = {
           'Endringslogger',
           'Felles',
           'Primitives',
+          'Experimental',
           'Altinn',
           'Avviklet',
         ],