+++ packages/react/src/components/Combobox/utils/defaultFilter.ts
@@ -1,0 +1,5 @@
+import { containsAllCharsInOrder } from '../../../utilities';
+import type { ComboboxFilter } from '../types/ComboboxFilter';
+
+export const defaultFilter: ComboboxFilter = (inputValue, options) =>
+  options.filter((option) => containsAllCharsInOrder(inputValue, option));